seo_agent_cli/core/tools/company_info_tool.py
from typing import Dict, Any
from pydantic import BaseModel, Field

# Pydanticモデルでツールの戻り値を定義
class CompanyInfo(BaseModel):
    id: str
    name: str = Field(description="会社名")
    business_description: str = Field(description="事業内容")
    target_audience: str = Field(description="ターゲット顧客層")

def get_company_info(company_id: str) -> CompanyInfo:
    """
    DBから指定されたIDの会社情報を取得する。
    
    :param company_id: 会社情報のID。
    :return: 会社情報。
    """
    print(f"--- TOOL: Executing get_company_info for ID: {company_id} ---")
    # TODO: データベースクライアントを呼び出すように実装する
    
    # ダミーデータの返却
    return CompanyInfo(
        id=company_id,
        name="株式会社サンプルテック",
        business_description="BtoB向けの最先端AIソリューションを提供しています。主な製品は、業務効率化SaaS「AutoTasker」です。",
        target_audience="従業員50名以上の中小企業から大企業の事業部長クラス。",
    )
